added custom timeline component of text generation and image generation
