+++ client/src/index.js
@@ -4,6 +4,7 @@ import {BrowserRouter, Routes, Route} from "react-router-dom"
 import './index.css';
 import Postanalysis from './pages/Postanalysis';
 import App from './App';
+import Generation from './components/Generation';
 
 
 export default function Approuter(){
@@ -12,6 +13,7 @@ export default function Approuter(){
     <Routes>
         <Route path="/" element={<App />}></Route>
         <Route path="/postanalysis" element={<Postanalysis />}></Route>
+        <Route path="/generate" element={<Generation />}></Route>
     </Routes>
     </BrowserRouter>
   )
